backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
prometheus_client
